# docs/05-api/openapi.yaml
openapi: 3.0.3
info:
  title: Inventory API
  version: 1.0.0
  description: API interna y pública para gestión de inventario StockFlow.
  contact:
    name: StockFlow Support
    email: soporte@tuempresa.com
servers:
  - url: https://inventario.tuempresa.com
    description: Production
  - url: http://localhost:8080
    description: Local backend
...
paths:
...
  /api/v1/users:
    get:
      tags: [Users]
      summary: Listar todos los usuarios
      description: Obtiene una lista paginada de todos los usuarios registrados en el sistema. (Solo administradores).
      operationId: listUsers
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/pageParam'
        - $ref: '#/components/parameters/sizeParam'
      responses:
        '200':
          description: Lista de usuarios recuperada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageUserResponse'
        '401':
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: Permisos insuficientes (Requiere rol ADMIN)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    post:
      tags: [Users]
      summary: Registrar nuevo usuario
      description: Registra un nuevo usuario en el sistema. (Solo administradores).
      operationId: createUser
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreateRequest'
      responses:
        '201':
          description: Usuario creado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Datos de entrada no válidos o conflicto de datos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: Email ya registrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/v1/users/{publicId}:
    put:
      tags: [Users]
      summary: Actualizar usuario existente
      description: Modifica los campos editables del perfil de un usuario.
      operationId: updateUser
      security:
        - bearerAuth: []
      parameters:
        - name: publicId
          in: path
          required: true
          description: Identificador público UUID del usuario.
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
      responses:
        '200':
          description: Usuario actualizado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          description: Usuario no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/v1/users/{publicId}/status:
    patch:
      tags: [Users]
      summary: Activar o desactivar usuario
      description: Habilita o deshabilita la cuenta de un usuario para permitir o denegar el acceso.
      operationId: toggleUserStatus
      security:
        - bearerAuth: []
      parameters:
        - name: publicId
          in: path
          required: true
          description: Identificador público UUID del usuario.
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [active]
              properties:
                active:
                  type: boolean
                  description: Nuevo estado lógico de activación.
                  example: false
      responses:
        '200':
          description: Estado del usuario actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          description: Usuario no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Introducir el Access Token JWT precedido del prefijo 'Bearer '.
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Clave de acceso estática para integradores externos autorizados.
  parameters:
    pageParam:
      name: page
      in: query
      description: Número de página a consultar (base 0).
      required: false
      schema:
        type: integer
        default: 0
    sizeParam:
      name: size
      in: query
      description: Cantidad máxima de registros por página.
      required: false
      schema:
        type: integer
        default: 10
  schemas:
...
    UserResponse:
      type: object
      properties:
        publicId:
          type: string
          format: uuid
          example: e2c92a95-5dbd-4a18-912c-354f9a0a03bb
        email:
          type: string
          format: email
          example: tecnico@tuempresa.com
        role:
          type: string
          enum: [ADMIN, TECNICO]
          example: TECNICO
        active:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
          example: "2026-05-22T08:00:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2026-05-22T10:30:00Z"
    UserCreateRequest:
      type: object
      required: [email, password, role]
      properties:
        email:
          type: string
          format: email
          example: nuevo.tecnico@tuempresa.com
        password:
          type: string
          minLength: 8
          example: tEcNiCo2026!
        role:
          type: string
          enum: [ADMIN, TECNICO]
          example: TECNICO
    UserUpdateRequest:
      type: object
      required: [email, role]
      properties:
        email:
          type: string
          format: email
          example: modificado@tuempresa.com
        role:
          type: string
          enum: [ADMIN, TECNICO]
          example: ADMIN
...
    ApiError:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: BAD_REQUEST
            message:
              type: string
              example: La justificación es obligatoria para dar de baja un equipo.
        timestamp:
          type: string
          format: date-time
          example: "2026-05-22T08:30:01Z"
    PageUserResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/UserResponse'
        pageNumber:
          type: integer
          example: 0
        pageSize:
          type: integer
          example: 10
        totalElements:
          type: integer
          example: 5
        totalPages:
          type: integer
          example: 1
        last:
          type: boolean
          example: true
...